requirements.txt
Django
psycopg2-binary==2.9.9
djangorestframework==3.14.0
python-decouple==3.8
